Fix Linux build issues and add install instructions for the same

+++ Source/src/dataanalyzer.cpp
@@ -225,11 +225,11 @@ void DataAnalyzer::run() {
 						break;
 					case Dso::WINDOW_BARTLETT:
 						for(unsigned int windowPosition = 0; windowPosition < this->lastBufferSize; windowPosition++)
-							*(this->window + windowPosition) = 2.0 / windowEnd * (windowEnd / 2 - abs(windowPosition - windowEnd / 2));
+							*(this->window + windowPosition) = 2.0 / windowEnd * (windowEnd / 2 - abs((int)(windowPosition - windowEnd / 2)));
 						break;
 					case Dso::WINDOW_TRIANGULAR:
 						for(unsigned int windowPosition = 0; windowPosition < this->lastBufferSize; windowPosition++)
-							*(this->window + windowPosition) = 2.0 / this->lastBufferSize * (this->lastBufferSize / 2 - abs(windowPosition - windowEnd / 2));
+							*(this->window + windowPosition) = 2.0 / this->lastBufferSize * (this->lastBufferSize / 2 - abs((int)(windowPosition - windowEnd / 2)));
 						break;
 					case Dso::WINDOW_GAUSS:
 						{
